requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.7.1=py37_0
astor=0.8.0=py37_0
blas=1.0=mkl
c-ares=1.15.0=h7b6447c_1001
cairo=1.14.12=h8948797_3
certifi=2019.9.11=py37_0
cffi=1.12.3=py37h2e261b9_0
cudatoolkit=10.0.130=0
cudnn=7.6.0=cuda10.0_0
cupti=10.0.130=0
cycler
dbus=1.13.6=h746ee38_0
expat=2.2.6=he6710b0_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut
freetype=2.9.1=h8a8886c_1
gast=0.2.2=py37_0
glib=2.56.2=hd408876_0
google-pasta=0.1.7=py_0
graphite2
grpcio=1.16.1=py37hf8bcb03_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.8.0=py37h989c5e5_3
harfbuzz
hdf5
icu=58.2=h211956c_0
intel-openmp
jasper=2.0.14=h07fcdf6_1
joblib=0.13.2=py37_0
jpeg=9b=habf39ab_1
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libglu=9.0.0
libopencv=3.4.2=hb342d67_1
libopus=1.3=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.8.0=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
markdown=3.1.1=py37_0
matplotlib=3.1.1=py37h5429711_0
mkl
mkl-service
mkl_fft
mkl_random
ncurses
ninja
numpy
numpy-base
opencv=3.4.2
openssl
pandas=0.25.1
pcre=8.43
pixman
protobuf=3.8.0
py-opencv=3.4.2
pycparser=2.19
pyparsing=2.4.2
pyqt=5.9.2
python-dateutil
pytz=2019.2
qt=5.9.7
readline=7.0=h7b6447c_5
scikit-learn=0.21.2=py37hd81dba3_0
scipy=1.3.1=py37h7c811a0_0
setuptools=41.0.1=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.12.0=py37_0
sqlite=3.29.0=h7b6447c_0
tensorboard=1.14.0
tensorflow=1.14.0
tensorflow-base=1.14.0
tensorflow-estimator=1.14.0
tensorflow-gpu=1.14.0
termcolor=1.1.0=py37_1
tk=8.6.8=hbc83047_0
tornado
werkzeug=0.15.5=py_0
wheel=0.33.4=py37_0
wrapt=1.11.2=py37h7b6447c_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
